+++ Src/app_i2c_lcd.c
@@ -470,7 +470,7 @@ void Show_Box_ID(uint16_t bID){
 	Lcd_Send_String((char*)strBoxID);
 }
 
-void Show_KWH(uint32_t ws){
+void Show_KWH(uint64_t ws){
 	char strTotalKWH[16];
 	static uint32_t preKwh = 0;
 	uint32_t kwh;
